requirements.txt
kuzu>=0.4.0
chromadb>=0.4.22
watchdog>=4.0.0
streamlit==1.31.0
pyyaml==6.0.1
requests==2.31.0
pandas>=2.2.0
plotly==5.18.0
networkx==3.2.1
matplotlib==3.8.2
openai>=1.0.0
python-dotenv==1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
tqdm>=4.66.0
python-multipart>=0.0.9

